Find whether the bill is splitted fairely or not.
Inputs Array of integers corresponding to the ordered items price, K the index of the item, one party refuse to eat.

Solution Calculate the total amount excluding the price in the specified index and divide by 2.

That is the amount one party should return to other party.

If nothing to return the print "Bon Appetit"