Find number of pairs in an array. The array will contain integers which may repeat. Find total number of value pairs from the array.
Solution:
Loop through each element in the array, and add each value into a map, increment for each occurence.
Same time keep pairCount variable to track the paris count, which will help to overcome the repeated looping.

In each iteration check the value corresponding to the value even or odd. If the value is even, then we can increment the pairCount value with one.